+++ src/constants/index.ts
@@ -1,0 +1,1 @@
+export {KEY} from "./keys.constant";
